PythonProject/mcp_server/requirements.txt
mcp
python-dotenv
google-generativeai